src/requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-cors-headers==3.13.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
django-phonenumber-field==6.3.0
phonenumbers==8.12.13
djangorestframework-simplejwt==5.2.0
requests==2.25.1
django-filter==21.1
drf-writable-nested==0.7.0
certifi==2022.9.14
chardet==4.0.0
idna==2.10
jira==3.4.1
oauthlib==3.2.1
PyJWT==2.5.0
pyparsing==3.0.9
packaging==21.3
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
typing_extensions==4.3.0
urllib3==1.26.12
django-apscheduler==0.6.2
drf-yasg==1.21.5
xero-python==1.23.0
Authlib==1.2.0
pyxero==0.9.3
django-storages==1.13.2
boto3==1.26.113
slack-sdk==3.21.2
django-sslserver==0.22
canvas==0.3
reportlab==4.0.0
pymsteams==0.2.2
msal==1.22.0
dnspython==2.3.0
celery>=5.2.3,<6.0
django_celery_beat==2.5.0
django_celery_results==2.5.1
channels==4.0.0
channels-redis==4.1.0
coverage==7.3.3
pytest==7.4.3
pytest-django==4.7.0
django-coverage==1.2.4
daphne==4.0.0